decomp3: spawn `target`, add merc and particle buckets and some temporary hacks (#3445)
This includes all the collision stuff needed to spawn `target`,
decompiles the sparticle code and adds some of the PC hacks needed for
merc to run (it doesn't work quite right and looks bad, likely due to a
combination of code copied from Jak 2 and the time of day hacks).

There are a bunch of temporary hacks (see commits) in place to prevent
the game from crashing quite as much, but it is still extremely prone to
doing so due to lots of missing functions/potentially bad decomp.

---------

+++ decompiler/analysis/find_skelgroups.cpp
@@ -570,7 +570,7 @@ void inspect_cloth_data_jak3(LetElement* let, DefskelgroupElement::StaticInfo& i
         auto array_set = dynamic_cast<SetFormFormElement*>(when_body->at(0));
         if (array_set) {
           // get elements
-          auto elts = when_body->elts().size() - 2;
+          auto elts = (int)when_body->elts().size() - 2;
           for (int i = 0; i < elts; i++) {
             auto parms_form = dynamic_cast<SetFormFormElement*>(when_body->at(i + 1));
             if (parms_form) {
